Add files via upload
This a markerless template matching tracker that is based on normalized-cross-correlation and uses the Gaussian distribution to estimate subpixel accuracy, which provides an accuracy of 0.01 pixel.  It also can be used with a set of images or videos. it can update template shape and color gradually. It's fast enough to be used for online tracking purposes.